+++ src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/get/GetUserDto.java
@@ -100,5 +100,5 @@ public class GetUserDto {
     /**
      * Identificador de la plataforma telegram utilizada en notificaciones.
      */
-    private Integer telegram_id;
+    private Long telegram_id;
 }
